requirements.txt
requests==2.32.4
pandas==2.3.0
matplotlib==3.10.3
seaborn==0.13.2
openai==1.93.0
python-dotenv==1.1.1
tqdm==4.67.1
google-generativeai==0.8.3 